Add support for zooming with standard ctrl+wheel
Now font size can be temporarily changed with usual Windows gesture -
Ctrl+Mouse Wheel.

This fix is discussed on Sourceforge in [ 1733641 ] Hotkey to resize fonts.

The calculation of GetMaxRect is modified to allow zooming beyond
main-screen limit, imposed by the design of cmd.exe. When the limit was
enforced, Console did not calculate the new buffer size correctly.