Fix monitoring of parent process
Unfortunately, there appears to be no way to monitor a process handle of
another user without special permissions.

The idea is: create a mutex that will be abandoned when Console terminates
unexpectedly. When that happens, the child knows that it's time to exit.

There is a problem: the watchdog mutex better be created in MainFrame or
other single-instance class, but SharedMemNames define wformat, so it can
be included only in one file (otherwise, linker complains).