Attempt to avoid funny selection with a "slow double-click"
The problem seems to exist due to an UpdateSelection(point), called from
UpdateConsoleView, and/or due to a slight mouse move between two clicks of
a double click, causing UpdateSelection(point) to change m_coordCurrent.

It looks like we can trick SelectionHandler and ConsoleView to behave if
we set the state above selstateSelecting. However selstateSelected is used
to protect from accidental "second copy", hence it can't be used.